feat: close P0-P2 gaps across all 9 pillars (332 → 408 tests)
P0 — 10 anonymization levels (was binary detection only):
  - New anonymize.rs: 10 levels from Highlight to FullRedaction
  - Levels 1-6 fully working, 7-10 basic (need QRNG/dataset for production)
  - 22 new tests

P0 — Q-AI LLM backend + prompt injection scanner:
  - prompt_guard.rs: 18 regex patterns, PromptGuardResult
  - ModelProvider enum: Claude/OpenAI/Local/Mock with API key routing
  - guarded_chat_sync() blocks unsafe prompts before dispatch
  - 22 new tests (14 guard + 8 routing)

P1 — Messenger persistence:
  - message_store.rs: MessageStore trait, InMemory + FileMessageStore (AES-256-GCM)
  - Ratchet integration: encrypt_and_store(), retrieve_and_decrypt()
  - 8 new tests including encrypted round-trip

P1 — Email SMTP + self-destruct:
  - email_transport.rs: PqcSmtpClient (RFC 5322, PQC headers, lettre behind feature flag)
  - SelfDestructTimer with Delete/Overwrite3Pass/OverwriteRandom
  - 5 new tests

P1 — VoIP session manager:
  - voip_session.rs: SDP offer/answer with PQ-KEM extensions
  - Directional SRTP keys via derive_srtp_keys_labeled()
  - Full handshake: create_offer → accept_offer → complete_handshake
  - 7 new tests (both sides derive matching shared secret)

P2 — Clippy: 13 warnings fixed across core + nist (0 warnings under -D)
P2 — Q-Mesh: MeshProvisioner + 44 tests (was 0), ESP32 JSON provisioning
P2 — FEATURES.md: honest code-verified percentages (avg 79% → 61%)

Test count: 332 → 408 passed, 0 failed

+++ browser/src-tauri/src/privacy/password_manager.rs
@@ -330,7 +330,7 @@ impl PasswordVault {
             let byte = self.entropy.read_bytes(1)[0] as usize;
             // Rejection sampling to avoid modulo bias.
             let limit = 256 - (256 % charset_len);
-            if (byte as usize) < limit {
+            if byte < limit {
                 out.push(CHARSET[byte % charset_len]);
             }
             attempts += 1;
